Defer http server startup
Prior to this commit, Micronaut was eagerly starting its embedded netty
server and it appeared there was no way to stop it from doing so.
Further research showed that this is easy, and being able to defer it
until everything else is starting up helps clean up a number of other
aspects like when the data dir is initialized.