Add --base-data-dir cli opt
And rename Options.{userDataDir => baseDataDir} to better convey this
option's nature. It defaults to the user's application data directory,
but its function is to serve as the base directory for the actual
application directory, and that can be customized to point somewhere
other than the system default user data directory. This is especially
useful to do during development and testing.