Remove --app-name|--conf mutual exclusivity constraint
The previous implementation was incorrect as it put the constraint on
--base-data-dir instead of --conf; it should have been between
--app-name and --conf. The point is moot now, however, as necessary
handling has been put in place. Basically, we need the value of
<app-name> early in order to construct the path to <app-data-dir>, but
this value may get overridden if the user's <conf> file contains an
entry for it. There is now a "2nd pass" of assigning this value from the
cli options to ensure that it takes precedence over any conf file value.