Add wip cli subcommands and dca bot logic
The new CLI subcommands here have been very quickly added in order to
exercise their respective endpoint operations. The actual naming and
structure of the subcommands will eventually resemble something more
like git's subcommand model. Likewise, the handling of the subcommands
is naive and cut-and-paste just for the time being.