Polish formatting and imports
This commit adds a .idea/codeStyles/Project.xml configured to avoid
import on demand for individual classes. The value is set to 99 because
that is the highest value allowed, but its meaning is "never".

The threshold for using .* on static imports remains at the default
value of 3.

These settings had been customized in my own global IDE settings,
but do not carry through to new users checking out the project, thus
adding the custom project-level configuration here.