Rename OfferRepository => Offerbook
The component now named Offerbook is essentially a DDD-style repository,
but it's worth breaking from the conventional *Repository naming here.

The main reason why is that this component is not just an implementation
detail; it is not just part of Bisq's "persistence layer" like such
components often are in traditional codebases. Rather, this component
maps directly onto an essential concept within Bisq's core domain of
bitcoin-to-fiat exchange. Bisq's order book (henceforth "orderbook") is
a term that is exposed in the UI and that users actually talk about in
natural/everyday language. People literally refer to Bisq's "offerbook"
and they would not know what you were talking about if you said Bisq's
"offer repository".

The notion of an offerbook is also key to distinguishing Bisq's
decentralized approach vs. the centralized approach of typical exchange
services.

Centralized exchanges typically operate on an "order book" model, where
users "place orders" to buy and sell, and the centralized service does
"automatic order matching" to indirectly (and usually fractionally)
connect buyers and sellers. The buyers and sellers are never aware of
each other in this model and everything happens through the exchange and
its order book and concomitant matching algorithms.

Bisq on the other hand has no "order book" because it has no "orders"
per se, and doesn't do any sort of "automatic order matching". Rather,
Bisq affords users the opportunity to _offer_ to buy or sell bitcoin
under certain terms, and affords would-be counterparties the opportunity
to discover those offers and take them. From that point forward, the
users are aware of one another and work together to complete their
desired exchange. Unlike a centralized "order book", Bisq's offerbook is
not in the middle of the exchange. It is rather just a tool, like a
community bulletin board, that people having a common interest use to
find each other.

This terminology, then, of an "offerbook", prominently and consistently
used, can help people in their journey of understanding the key
differences between Bisq's decentralized, human-centric model and
ordinary exchanges' centralized, algorithm-centric model.

As a technical component, Bisq's offerbook is similar in a number ways
to bitcoin's so-called "mempool". All bitcoin nodes maintain a "memory
pool of pending transactions" known as a "mempool". Each node's mempool
is likely to be somewhat different than that of the next, but all are
sufficiently consistent with each another that a kind of global,
decentralized entity that one might refer to as *the* bitcoin mempool
emerges in the user's mind. It is technically incorrect to refer to it
this way, because indeed there is no singular mempool manifest anywhere
in the system, but it is nevertheless a helpful mental model, and one
that, at a high enough level of abstraction, ceases to be incorrect.
After all, bitcoin a as global system *does* maintain a pool of pending
transactions; the implementation of that pool happens to be
decentralized and peer-to-peer. But that hardly matters in practice for
the everyday user. All they know is whether their bitcoin transaction
has been confirmed or that it is still pending. In any case, bitcoin's
mempool(s) have now become a widely understood and talked about part of
the bitcoin ecosystem, even amongst many non-technical users.

And like bitcoin's mempool component, Bisq's offerbook is a distributed,
in-memory collection of all offers to exchange that have not yet been
taken. Each Bisq node maintains its own independent offerbook, but at
system-wide scale, what emerges is something that looks and behaves like
a singular, unified offerbook. Given these similarities, much of the
same mental modeling and reasoning that users do about bitcoin's mempool
can and should be done about Bisq's offerbook. So it's critical that we
name it correctly and give it its proper place as a first class
component, not just within the UI, but in in the architecture of and
ubiquitous language [1] around Bisq.

[1]: https://martinfowler.com/bliki/UbiquitousLanguage.html