Refine project-level README
This commit restores the readme to a basic outline to be filled in with
prose and links to other documents later. The previous content was a
work in progress, and while much of it is valuable and may be brought
back, much of it was also incomplte and would have been distracting or
confusing for anyone coming fresh into the project.