Apply license header to existing sources
With the previous commits, all newly-created java source files will
automatically have the Bisq license header applied to them (when they
are created in IDEA, anyway). This commit applies the license header to
all existing sources.

The following script was used to apply this change:

    for sourcefile in $(git ls-files | grep java$ | grep -v ^client |  grep -v \.idea);
        do
            cat .idea/fileTemplates/includes/File\ Header.java $sourcefile > $sourcefile.tmp;
            mv $sourcefile.tmp $sourcefile;
        done

Note that the generated sources in the `client` module were excluded
from having the license file header applied, as this change would get
wiped out on subsequent generation runs and cause spurious diffs. It
would be possible to configure client generation to include the headers
with a custom template per https://openapi-generator.tech/docs/customization/,
but this not likely to be worth the effort any time soon.