Rename :client => :clients
This is done in anticipation of the addition of kotlin and python
clients alongside the existing java client, and is done for symmetry
with the top-level 'apps', 'demos' and 'docs' directories. Their plural
form signals that there are multiple apps, demos, docs and indeed now
clients within them. Contrast these with the singular naming and nature
of the 'core' directory and its contents.