Add initial generation of kotlin client
This work in progress commit adds the result of running
openapi-generator with the following parameters from the root project
directory:

    openapi-generator generate \
        -o clients/kotlin \
        -g kotlin \
        -i core/oas/build/classes/java/main/META-INF/swagger/bisq-openapi-2.1.0.yml

The output is being committed as-is after that generation run, with
the exception of removing generated gradlew scripts and binaries with
the following command run within the new clients/kotlin directory:

    git rm -rf gradle* settings.gradle

The only other modification in this commit is adding the :clients:kotlin
module to project's root level settings.gradle.kts file.

The build does not run successfully at this point. Further modifications
to the generated build file (including migrating build.gradle =>
build.gradle.kts will resolve this. The goal will be to have the build
file and various openapi-generator configuration files mirror as closely
as possible those in the existing :clients:java module.