Rename Info => NodeInfo
This change has been planned for a while, but doing it now, as part of
adding the new kotlin client in order to demonstrate the effect of
core API changes bubbling up through to changes in the client libraries
as part of the openapi client generation process. This has been working
for some time with :clients:java, and it now works for :clients:kotlin
as well.