Rename java client 'invokerPackage' to bisq.client.java{=>.infrastructure}
This is for symmetry with the newly added kotlin client, whose
'infrastrucutre' package is hard-coded and non-configurable. On the
kotlin side, for example, the generated `ApiClient` class is packaged as

    bisq.client.kotlin.infrastructure.ApiClient

prior to this change, the java client's equivalent class was packaged as

    bisq.client.java.ApiClient

but is now packaged as

    bisq.client.java.infrastructure.ApiClient

Nothing changes functionally; this is just for consistency across the
client library implementations.

Note: unfortunately, the 'auth' package cannot be named consistently
across the java and kotlin client libraries. Again, due to hard-coding,
the kotlin client generator forces the auth package to live directly
under the base package at `${packageName}.auth`, while the java client
generator always places it at `${invokerPackage}.auth`. This means there
is no way to have identical packaging for both the infrastructure and
auth packages across the two implementations. A minor nit, but worth
mentioning should someone notice and try to fix it in the future.

The Structure101 configuration file has been updated to mimic what it
would be like if kotlin's auth package were nested within its
infrastructure package. So at least when looking at s101, this symmetry
does appear to be present even though in the codebase it is not.